Move InspectorPanel struct to the header file
This patch moves the InspectorPanel struct to the header file
winspector.h. Now we can use the InspectorPanel in other files.
