+++ src/winspector.h
@@ -23,6 +23,62 @@
 #define WINSPECTOR_H_
 
 #include "window.h"
+#include "config.h"
+
+typedef struct InspectorPanel {
+	struct InspectorPanel *nextPtr;
+
+	WWindow *frame;
+	WWindow *inspected;	/* the window that's being inspected */
+	WMWindow *wwin;
+	Window parent;
+	char *title;		/* InspectorPanel title */
+
+	/* common stuff */
+	WMButton *revertBtn;
+	WMButton *applyBtn;
+	WMButton *saveBtn;
+	WMPopUpButton *pagePopUp;
+
+	/* first page. general stuff */
+	WMFrame *specFrm;
+	WMButton *instRb;
+	WMButton *clsRb;
+	WMButton *bothRb;
+	WMButton *defaultRb;
+	WMButton *selWinB;
+	WMLabel *specLbl;
+
+	/* second page. attributes */
+	WMFrame *attrFrm;
+	WMButton *attrChk[11];
+
+	/* 3rd page. more attributes */
+	WMFrame *moreFrm;
+#ifndef XKB_BUTTON_HINT
+	WMButton *moreChk[11];
+#else
+	WMButton *moreChk[12];
+#endif
+
+	/* 4th page. icon and workspace */
+	WMFrame *iconFrm;
+	WMLabel *iconLbl;
+	WMLabel *fileLbl;
+	WMTextField *fileText;
+	WMButton *alwChk;
+	WMButton *browseIconBtn;
+	WMFrame *wsFrm;
+	WMPopUpButton *wsP;
+
+	/* 5th page. application wide attributes */
+	WMFrame *appFrm;
+	WMButton *appChk[3];
+
+	unsigned int done:1;
+	unsigned int destroyed:1;
+	unsigned int choosingIcon:1;
+} InspectorPanel;
 
 void winspector_destroy(struct InspectorPanel *panel);
 void wShowInspectorForWindow(WWindow *wwin);
